src/mod-4-multi-agent/crewai/requirements.txt
crewai==0.119.0
crewai-tools==0.62.3

psycopg2-binary==2.9.9
sqlalchemy==2.0.35

apscheduler==3.10.4

python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

faker==28.4.1

langfuse==2.58.2
openlit==1.28.1
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2

python-dateutil==2.9.0
pytz==2024.2

colorlog==6.8.2